requirements.txt
setuptools==46.1.3
asgiref==3.2.3
Django==3.0.2
gunicorn==20.0.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
tensorflow-cpu==2.1.0
matplotlib==3.1.1
tqdm==4.36.1
ipython==7.12.0
imageio==2.6.1
Pillow==6.2.1